+++ controllers/api/v1/user_controller/update_email.go
@@ -4,7 +4,7 @@ import (
 	"flop/helper/api_response_helper"
 	"flop/middleware"
 	"flop/models/database_model"
-	"flop/repositories/users"
+	"flop/repositories/users_repository"
 	jwt "github.com/appleboy/gin-jwt/v2"
 	"github.com/gin-gonic/gin"
 	"net/http"
@@ -27,7 +27,7 @@ func UpdateEmail(c *gin.Context, authMiddleware *jwt.GinJWTMiddleware) {
 	newEmail := c.Request.FormValue("new_email")
 
 	user := database_model.Users{}
-	users.UpdateUserEmail(&user, currentEmail, newEmail)
+	users_repository.UpdateUserEmail(&user, currentEmail, newEmail)
 
 	newJWT, _, err := authMiddleware.TokenGenerator(&user)
