+++ controllers/api/v1/user_controller/user_info.go
@@ -2,7 +2,7 @@ package user_controller
 
 import (
 	"flop/middleware"
-	"flop/repositories/users"
+	"flop/repositories/users_repository"
 	"github.com/gin-gonic/gin"
 	"net/http"
 )
@@ -19,6 +19,6 @@ import (
 //	@Security		ApiKeyAuth
 func GetUserInfo(c *gin.Context) {
 	userId := middleware.GetUserIdFromJWT(c)
-	currentUser := users.GetOneUserById(userId)
+	currentUser := users_repository.GetOneUserById(userId)
 	c.IndentedJSON(http.StatusOK, currentUser)
 }
