--- repositories/user_logged_in_devices/insert_on_sign_up.go
@@ -1,11 +1,0 @@
-package user_logged_in_devices
-
-import (
-	"flop/config/database"
-	database2 "flop/models/database_model"
-	"gorm.io/gorm"
-)
-
-func InsertOnSignUp(request *database2.UserLoggedInDevices) *gorm.DB {
-	return database.DB.Create(&request)
-}
