+++ repositories/user_logged_in_devices_repository/insert_on_sign_up.go
@@ -1,0 +1,11 @@
+package user_logged_in_devices_repository
+
+import (
+	"flop/config/database"
+	"flop/models/database_model"
+	"gorm.io/gorm"
+)
+
+func InsertOnSignUp(request *database_model.UserLoggedInDevices) *gorm.DB {
+	return database.DB.Create(&request)
+}
