--- repositories/users/get_user_by_id.go
@@ -1,12 +1,0 @@
-package users
-
-import (
-	"flop/config/database"
-	database2 "flop/models/database_model"
-)
-
-func GetOneUserById(userId uint) database2.Users {
-	var user database2.Users
-	database.DB.Where("id = ?", userId).First(&user)
-	return user
-}
